Basic view catalog GUI now working and shows images
- Added CatalogItemView which needed for the table view in catalog GUI
- Added imagePath to catalogItem and in entityFactory
- Added panter.jpg and buttercup.jpg as example images
- Made sure we set catalogGUI as the current UI in client before setting
the client in catalogGUI

+++ src/product/Product.java
@@ -23,7 +23,7 @@ public class Product implements Serializable
   	*  @param newType
   	*/
 //*************************************************************************************************
-	public Product(int newID, String newName, String newType,
+	public Product(long newID, String newName, String newType,
 					float productPrice, int productAmount,String productColor)
 	{
 		this.productID = newID;
